+++ src/todo/remove.js
@@ -1,0 +1,8 @@
+import {reloadTodoList, todoList} from "./common";
+
+function removeTodo(todoId) {
+    todoList.removeById(todoId)
+    reloadTodoList();
+}
+
+export {removeTodo};
